requirements.txt
pycups==2.0.4
PyPDF2==3.0.1
